Refactor after rebase to dev
